Fix compiler warnings for integer conversions and literals
Add an explicit static_cast to uint8_t in the evidence store integrity
check to resolve implicit conversion warnings caused by integer promotion
during XOR operations. Additionally, replace 'ull' suffixes with UINT64_C
macros in ABI policy tests for more portable 64-bit integer literals.

+++ src/evidence_store.cpp
@@ -988,7 +988,8 @@ bool EvidenceStore::deserialize(const std::vector<uint8_t> &blob,
   FactDigest expected = sha256_bytes(envelope);
   uint8_t difference = 0;
   for (size_t i = 0; i != kDigestSize; ++i)
-    difference |= expected.bytes[i] ^ blob[blob.size() - kDigestSize + i];
+    difference |= static_cast<uint8_t>(
+      expected.bytes[i] ^ blob[blob.size() - kDigestSize + i]);
   if (difference != 0)
     return fail(error, "evidence-store SHA-256 integrity check failed");
